+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-
 # Change Log
+
+## [1.0.1] - 2018-02-02
+- Fixed the display of null types in the report
+- Fixed the display of data types found in registries
+- Added validation of primitive types
+
+## [1.0.0] - 2018-01-26
+- Various bug fixes; getting into standard release cadence
 
 ## [0.9.0] - 2016-09-06
 - Initial Public Release